Fix file name of KeyStore, references to com.nguyenmp.grawler.com.nguyenmp.grawler.utils.*, and updated parsing for changes introduced during mid-Febuary downtime.
I had originally named the file GrawlerKeystore but I loaded it using GrawlerKeyStore.  This issue only arises when the user is on a Linux machine because filenames are case senstivie in that OS.  Additionally, I used forwards slashes instead of double backslashes since it caused Linux to not find the file.

I forgot to remove references to Utils.java under the oddly named package when I removed it from git.

During mid-February, GOLD's source code was modified slightly.  Some of the delimiters we used to parse the course data was also changed so we modified those delimiters appropriately.  BasicFindCourses, FinalExam, and Grawler have been fixed.